made changes to make build-env-endex.js for gcp integration
