+++ config/config.js
@@ -10,36 +10,7 @@ const secretVariable = {
     database: process.env.DB_DATABASE_NAME,
     dialect: process.env.DB_DIALECT,
     seederStorage: process.env.DB_SEEDER_STORAGE,
-    awsConfig: aws-config,
-    gcpConfig: gcp-config,
-    // awsCloudwatch: {
-    //     enableAwsLogger: process.env.AWS_LOG_ENABLE === 'true',
-    //     logGroupName: process.env.AWS_LOG_GROUP_NAME,
-    //     logStreamName: process.env.AWS_LOG_STREAM_NAME,
-    //     accessKeyId: process.env.AWS_LOG_ACCESS_KEY_ID,
-    //     secretKey: process.env.AWS_LOG_SECRET_KEY,
-    //     region: process.env.AWS_LOG_REGION
-    // },
-    // awsS3: {
-    //     access_key_id: process.env.AWS_S3_ACCESS_KEY_ID,
-    //     secret_access_key: process.env.AWS_S3_SECRET_ACCESS_KEY,
-    //     region: process.env.AWS_S3_REGION,
-    //     bucketName: process.env.AWS_S3_BUCKET_NAME
-    // },
-    // awsSQS: {
-    //     accessKeyId: process.env.AWS_SQS_ACCESS_KEY_ID,
-    //     accessKey: process.env.AWS_SQS_ACCESS_KEY,
-    //     region: process.env.AWS_SQS_REGION,
-    //     url: process.env.AWS_SQS_URL,
-    //     arn: process.env.AWS_SQS_ARN,
-    //     deadLetterQueue: process.env.AWS_SQS_DEAD_LETTER_QUEUE
-    // },
-    // gcp: {
-    //     projectName: process.env.GOOGLE_PROJECT_NAME,
-    //     keyFile: process.env.GOOGLE_APPLICATION_CREDENTIALS,
-    //     logStreamName: process.env.GOOGLE_LOG_STREAM_NAME,
-    //     enableGcpLogger: process.env.GOOGLE_LOG_ENABLE === 'true'
-    // },
+    cloudServiceProviderConfig,
     logType: process.env.LOG_TYPE,
     clsNameSpace: process.env.CLS_NAMESPACE,
     cronitorSecretKey: process.env.CRONITOR_SECRET_KEY,
